requirements.txt
flask==3.1.3
werkzeug==3.1.6
flask-sqlalchemy==3.1.1
authlib==1.6.9
requests==2.33.1
python-dotenv==1.2.2
pytest==8.3.5
pytest-flask==1.3.0
